+++ EssentialFeed/EssentialFeedTests/Feed Cache/XCTestCase+FeedStoreSpecs.swift
@@ -1,0 +1,67 @@
+//
+//  XCTestCase+FeedStoreSpecs.swift
+//  EssentialFeedTests
+//
+//  Created by Deniz Tutuncu on 8/22/21.
+//
+
+import XCTest
+import EssentialFeed
+
+extension FeedStoreSpecs where Self: XCTestCase {
+    @discardableResult
+    func insert(_ cache: (feed: [LocalFeedImage], timestamp: Date), to sut: FeedStore) -> Error? {
+        let exp = expectation(description: "Wait for cache retrieval")
+        var insertionError: Error?
+        
+        sut.insert(cache.feed, timestamp: cache.timestamp) { receivedInsertionError in
+            insertionError = receivedInsertionError
+            exp.fulfill()
+        }
+        wait(for: [exp], timeout: 1.0)
+        return insertionError
+    }
+    
+    @discardableResult
+    func deleteCache(from sut: FeedStore) -> Error? {
+        let exp = expectation(description: "Wait for cache deletion")
+        var deletionError: Error?
+        sut.deleteCachedFeed { receivedDeletionError in
+            deletionError = receivedDeletionError
+            exp.fulfill()
+        }
+        wait(for: [exp], timeout: 8.0)
+        return deletionError
+    }
+    
+    func expect(_ sut: FeedStore, toRetrieveTwice expectedResult: RetrieveCachedFeedResult, file: StaticString = #file, line: UInt = #line) {
+        expect(sut, toRetrieve: expectedResult)
+        expect(sut, toRetrieve: expectedResult)
+    }
+    
+    func expect(_ sut: FeedStore, toRetrieve expectedResult: RetrieveCachedFeedResult, file: StaticString = #file, line: UInt = #line) {
+        let exp = expectation(description: "Wait for cache retrieval")
+        
+        sut.retrieve { retrieveResult in
+            switch (expectedResult, retrieveResult) {
+            
+            case (.empty, .empty),
+                 (.failure, .failure):
+                break
+                
+            case let (
+                .found(expectedResultFeed, expectedResultTimestamp),
+                .found(retrieveRetrieveFeed, retrieveRetrieveTimestamp)
+            ):
+                
+                XCTAssertEqual(expectedResultFeed, retrieveRetrieveFeed)
+                XCTAssertEqual(expectedResultTimestamp, retrieveRetrieveTimestamp)
+                
+            default:
+                XCTFail("Expected to retrieve \(expectedResult), got \(retrieveResult) instead")
+            }
+            exp.fulfill()
+        }
+        wait(for: [exp], timeout: 3.0)
+    }
+}
